tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.interpolate.coloralpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.interpolate.coloralpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.interpolate.coloralpha</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var g = ctx.createLinearGradient(0, 0, 100, 0);
  g.addColorStop(0, 'rgba(255,255,0, 0)');
  g.addColorStop(1, 'rgba(0,0,255, 1)');
  ctx.fillStyle = g;
  ctx.fillRect(0, 0, 100, 50);
  _assertPixelApprox(canvas, 25,25, 190,190,65,65, 3);
  _assertPixelApprox(canvas, 50,25, 126,126,128,128, 3);
  _assertPixelApprox(canvas, 75,25, 62,62,192,192, 3);
  t.done();

});
</script>
